restructured api call
